+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineTests.cs
@@ -507,6 +507,13 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VirtualMachineGuestAttestation");
         }
+        
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVMandVMSSTimeCreated()
+        {
+            TestRunner.RunTestScript("Test-VMandVMSSTimeCreated");
+        }
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]