+++ src/Storage/Storage/Common/AzureStorageFileShare.cs
@@ -22,6 +22,9 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
     using global::Azure.Storage.Files.Shares;
     using global::Azure.Storage;
     using Microsoft.WindowsAzure.Commands.Storage.Common;
+    using global::Azure.Storage.Files.Shares.Models;
+    using Microsoft.Azure.Storage.Auth;
+    using System.Management.Automation;
 
     /// <summary>
     /// Azure storage file object
@@ -46,6 +49,11 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
         public bool IsSnapshot { get; private set; }
 
         /// <summary>
+        /// Gets a value indicating whether this share is deleted.
+        /// </summary>
+        public bool? IsDeleted { get; private set; }
+
+        /// <summary>
         /// file share last modified time
         /// </summary>
         [Ps1Xml(Label = "LastModified", Target = ViewControl.Table, ScriptBlock = "$_.LastModified.UtcDateTime.ToString(\"u\")", Position = 2, TableColumnWidth = 20)]
@@ -65,7 +73,14 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
             {
                 if (privateShareClient == null)
                 {
-                    privateShareClient = GetTrack2FileShareClient(this.CloudFileShare, (AzureStorageContext)this.Context);
+                    if (this.IsDeleted == null || !this.IsDeleted.Value)
+                    {
+                        privateShareClient = GetTrack2FileShareClient(this.CloudFileShare, (AzureStorageContext)this.Context);
+                    }
+                    else
+                    {
+                        throw new InvalidJobStateException("This share is already deleted. Can't create ShareClient for it.");
+                    }
                 }
                 return privateShareClient;
             }
@@ -89,6 +104,19 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
         private global::Azure.Storage.Files.Shares.Models.ShareProperties privateShareProperties = null;
 
         /// <summary>
+        /// XSCL Track2 File share List properties
+        /// </summary>
+        public global::Azure.Storage.Files.Shares.Models.ShareItem ListShareProperties { get; private set; }
+
+        /// <summary>
+        /// Delete Share version ID
+        /// </summary>
+        public string VersionId { get; }
+
+        private ShareClientOptions shareClientOptions { get; set; }
+        private ShareServiceClient shareService { get; set; }
+
+        /// <summary>
         /// Azure storage file constructor
         /// </summary>
         /// <param name="share">Cloud file share object</param>
@@ -104,8 +132,114 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
             Context = storageContext;
         }
 
+        /// <summary>
+        /// Azure storage file share constructor from Track2 get file properties output
+        /// </summary>
+        public AzureStorageFileShare(ShareClient shareClient, AzureStorageContext storageContext, ShareProperties shareProperties = null, ShareClientOptions clientOptions = null)
+        {
+            Name = shareClient.Name;
+            this.privateShareClient = shareClient;
+            CloudFileShare = GetTrack1FileShareClient(shareClient, storageContext.StorageAccount.Credentials);
+            if (shareProperties != null)
+            {
+                privateShareProperties = shareProperties;
+                LastModified = shareProperties.LastModified;
+                SnapshotTime = Util.GetSnapshotTimeFromUri(ShareClient.Uri);
+                if (SnapshotTime != null)
+                {
+                    IsSnapshot = true;
+                }
+                Quota = shareProperties.QuotaInGB;
+            }
+            Context = storageContext;
+            shareClientOptions = clientOptions;
+        }
+
+        /// <summary>
+        /// Azure storage file share constructor from Track2 list share output
+        /// </summary>
+        public AzureStorageFileShare(ShareClient shareClient, AzureStorageContext storageContext, ShareItem shareItem, ShareClientOptions clientOptions = null)
+        {
+            Name = shareClient.Name;
+            //this.privateShareClient = shareClient;
+            CloudFileShare = GetTrack1FileShareClient(shareClient, storageContext.StorageAccount.Credentials);
+            if (shareItem != null)
+            {
+                this.ListShareProperties = shareItem;
+                this.IsDeleted = shareItem.IsDeleted;
+                this.VersionId = shareItem.VersionId;
+                if (shareItem.Properties != null)
+                {
+                    privateShareProperties = shareItem.Properties;
+                    LastModified = shareItem.Properties.LastModified;
+                    Quota = shareItem.Properties.QuotaInGB;
+                }
+                if (!string.IsNullOrEmpty(shareItem.Snapshot))
+                {
+                    SnapshotTime = DateTimeOffset.Parse(shareItem.Snapshot).ToUniversalTime();
+                    IsSnapshot = true;
+                    this.privateShareClient = GetTrack2FileShareClient(CloudFileShare, storageContext, clientOptions);
+                    this.privateShareClient = this.privateShareClient.WithSnapshot(shareItem.Snapshot);
+                }
+            }
+            Context = storageContext;
+            shareClientOptions = clientOptions;
+        }
+
+        /// <summary>
+        /// Azure storage file share constructor from Track2 list share output, with deleted share
+        /// </summary>
+        public AzureStorageFileShare(ShareServiceClient shareService, string shareName, AzureStorageContext storageContext, ShareItem shareItem, ShareClientOptions clientOptions = null)
+        {
+            this.shareService = shareService;
+            this.Name = shareName;
+            if (shareItem != null)
+            {
+                this.ListShareProperties = shareItem;
+                this.IsDeleted = shareItem.IsDeleted;
+                this.VersionId = shareItem.VersionId;
+                if (shareItem.Properties != null)
+                {
+                    privateShareProperties = shareItem.Properties;
+                    LastModified = shareItem.Properties.LastModified;
+                    Quota = shareItem.Properties.QuotaInGB;
+                }
+                if (!string.IsNullOrEmpty(shareItem.Snapshot))
+                {
+                    SnapshotTime = DateTimeOffset.Parse(shareItem.Snapshot).ToUniversalTime();
+                    IsSnapshot = true;
+                }
+            }
+            Context = storageContext;
+            shareClientOptions = clientOptions;
+        }
+
+        // undelete a soft deleted share
+        public void UndeleteShare()
+        {
+            if (this.IsDeleted != null && this.IsDeleted.Value)
+            {
+                this.shareService.UndeleteShare(this.Name, this.VersionId);
+            }
+            else
+            {
+                throw new InvalidJobStateException("This share is not soft deleted, so can't undelete it.");
+            }
+        }
+
+        // Convert Track2 File share object to Track 1 file share object
+        public static CloudFileShare GetTrack1FileShareClient(ShareClient shareClient, StorageCredentials credentials)
+        {
+            if (credentials.IsSAS) // the Uri already contains credentail.
+            {
+                credentials = null;
+            }
+            CloudFileShare track1CloudFileShare = new CloudFileShare(shareClient.Uri, credentials);
+            return track1CloudFileShare;
+        }
+
         // Convert Track1 File share object to Track 2 file share Client
-        protected static ShareClient GetTrack2FileShareClient(CloudFileShare cloudFileShare, AzureStorageContext context)
+        public static ShareClient GetTrack2FileShareClient(CloudFileShare cloudFileShare, AzureStorageContext context, ShareClientOptions clientOptions = null)
         {
             ShareClient fileShareClient;
             if (cloudFileShare.ServiceClient.Credentials.IsSAS) //SAS
@@ -121,16 +255,16 @@ namespace Microsoft.WindowsAzure.Commands.Common.Storage.ResourceModel
                 {
                     fullUri = fullUri + "?" + sas;
                 }
-                fileShareClient = new ShareClient(new Uri(fullUri));
+                fileShareClient = new ShareClient(new Uri(fullUri), clientOptions);
             }
             else if (cloudFileShare.ServiceClient.Credentials.IsSharedKey) //Shared Key
             {
                 fileShareClient = new ShareClient(cloudFileShare.SnapshotQualifiedUri,
-                    new StorageSharedKeyCredential(context.StorageAccountName, cloudFileShare.ServiceClient.Credentials.ExportBase64EncodedKey()));
+                    new StorageSharedKeyCredential(context.StorageAccountName, cloudFileShare.ServiceClient.Credentials.ExportBase64EncodedKey()), clientOptions);
             }
             else //Anonymous
             {
-                fileShareClient = new ShareClient(cloudFileShare.SnapshotQualifiedUri);
+                fileShareClient = new ShareClient(cloudFileShare.SnapshotQualifiedUri, clientOptions);
             }
 
             return fileShareClient;